remove file duplicates
